test: disable cabal-fmt for primer-service
Unfortunately cabal-fmt does not yet support references to cabal
sublibraries, breaking the source code checks. Support has been
implemented, so hopefully we will be able to revert this commit when the
next release of cabal-fmt is made.

(This commit is preparation for adding some tests to primer-service that
will require primer:primer-hedgehog.)