chore: let HLS automatically work out the cradle
Since we are using an auto-generated (by implicit-hie/gen-hie) hie.yaml,
we may as well let HLS work out the cradle by itself -- it has been
pretty good at this (outside of esoteric setups) for a few years now.
This change will ensure it never gets out of sync (for instance, when we
create a new package in the project).