fix: ensure ToSchema is derived in same way as ToJSON
Currently PrimerJSON is simply the default encoding, so our instances
are consistent. However, there is a latent bug which will be triggered
if we change the options -- for some types the ToJSON instance will
change but the ToSchema one will not.

We ensure that they are consistent under all options, by deriving
ToSchema via PrimerJSON whenever we do so for ToJSON.