test: automatic validation ToJSON/ToSchema for entire api
We use validateEveryToJSON to QuickCheck every type appearing in our
OpenAPI. Since this uses on QuickCheck, we pull in some extra
dependencies and write some orphan instances wrapping our hedgehog
generators. Unfortunately this does not preserve shrinking.

Note that this does not currently test any more types than our
hand-enumerated tests do. However, when we extend the API in future,
the compiler will check that we have added the appropriate Arbitrary
instances. (It is probably still worth manually adding hedgehog tests
for any new types added, since (a) they will have better shrinking and
(b) it is easy to do, especially since we will implement the hedgehog
generator anyway, to be wrapped in an Arbitrary instance.)