refactor: rethink OperationId implementation
It is now implemented similar to Description, used as
`OperationId "..." :> api`, rather than directly wrapping an endpoint.
All operation ids in a path to an endpoint are concatenated.

This style integrates better with tools that consume a servant api, in
particular automatic openapi validation of all types in the api. (This
will be added in a subsequent commit.)