fix: should not offer case on primitives
It is a bit awkward to plumb around a TypeDefMap. An alternative is
(as mentioned in the previous commit) to declare the module "Primitives"
as special (and forbid users defining their own module called that).
Then we could detect whether a type is a primitive or an ADT by name
without needing a TypeDefMap.