fix: copy/paste scoping for shadowing in inputs
We now take care to handle the case where there are binders in the input
which shadow each other, rather than naively comparing names. Note that
identity of binders is controlled by ID, whereas shadowing is controlled
by Name; whilst there is a mapping ID -> Name, it is not injective.