refactor: split Primitives to improve build parallelism
We split the definition of the PrimDef enumeration to its own module.
This is the bulk of the compile time for this module, because of the
derived instances. The definition itself has few dependencies, but the
Primitives module is a minor bottleneck in the build, so this split
should increase parallelism.