refactor: extract progAllDefs helper
This returns all defs along with whether they are mutable (in a "home"
module) or immutable (in an imported module). This information is
currently unused, but will be important in a subsequent commit.

We also rename the constructors of ModuleLookup (an internal helper) to
free up the nice name 'Editable' for the (exported) information that
'progAllDefs' returns.