chore: these tests are not utilities
Somehow we have some tests in the "Utilities" section of the module. We
move them to the main section.

I suspect they were simply added after the last unit test in the module,
without realising that test is part of the utilities (it checks that our
utility defaultFullProg is sensible).