refactor: rename TDIUnknownADT as it may be a primitive
Since we cannot know whether an unknown thing is an algebraic data type
or an in-built primitive, we shouldn't call them both TDIUnknownADT, as
that ADT suffix is misleading.

One option for future work would be to bless a set of primitive modules
and then we could distinguish an unknown ADT from a primitive based on
name.