feat: type level lets (core & utils only)
Similarly to 'lettype', these will be unsupported in the typechecker. It
is expected that (for now) they are only ephemeral nodes used during
evaluation. However, this commit does not implement support in the
evaluators or generators. (It does implement support everywhere other
than the typechecker, the evaluators and the generators.)