fix: don't unconditionally remove database ops from queue in `serve`.
Thus far, we've only been concerned about handling `Rel8DbException`s,
which we assume may indicate a problem with an individual database
operation, but not the database in general; e.g., perhaps a student's
program didn't get migrated correctly. Therefore, we've simply logged
the exception and continued serving operations, which implicitly
discards the failed operation, because prior to this change, `serve`
removed operations from the op queue before running the actual
database operation, and did not put them back on the queue in the
event that they failed.

However, as described in
https://github.com/hackworthltd/primer/issues/674, soon we'll need to
handle exceptions where the database operation should be retried, not
discarded. For example, if a database update fails because the
connection went down temporarily, we don't want to discard that
database update: there is nothing wrong with the update itself, it
just failed due to a temporary condition.

This commit changes the behavior of `serve` and the `primer-service`
main program such that `serve` no longer removes ops from the database
queue until they run to completion with no exceptions. Now when a
database operation throws an exception, then when the main program
handles the exception, the op which generated the exception will still
be on the database op queue; and the main program can decide to
proceed, and whether to retry or discard the failing operation,
depending on the sort of exception that occurred.

Note that this commit does not actually fix
https://github.com/hackworthltd/primer/issues/674, because we still
don't handle database connection exceptions properly. However, it does
lay the groundwork for fixing that issue in a subsequent commit.
Despite the function-level changes and refactored responsibility that
this commit introduces, the big-picture functionality hasn't actually
changed.