fix: typed generator shadowing issue
It is possible for the typed generator to produce a context akin to
  foo (type var) : TYPE, foo (term var) : foo
breaking the invariant that we never shadow type variables appearing
free in the context-assigned types (here the term variable 'foo' shadows
the type variable 'foo' appearing in the type of the term variable 'foo'
(i.e. itself)). This can lead to generating terms that are ill-scoped.

This led to test suite failures:
  cabal run primer-test -- -p 'smartholes idempotent chk' --hedgehog-replay "Size 35 Seed 1731114382788457852 2155585848329412751"
  cabal run primer-test -- -p 'tcWholeProg idempotent' --hedgehog-replay "Size 23 Seed 5931353374268303457 14072037541920620699"