feat: Initial logging support.
As we expect `primer-service` to run in some sort of container
service, we're content for now just to log to stdout and let the
container orchestration system handle shipping logs. This initial
implementation does just that: output log messages to stdout with a
severity tag.

In this commit, we add logging of database issues. Logging of "core"
and API-related issues is left for further work.

Note that this commit drops the monomorphized `runRel8Db` runner,
because it would require that we choose a specific logging
implementation, and that doesn't seem in scope for a library function.