feat: typed generator produces empty lists less often
We change the frequency of producing empty lists of definitions or
typedefs to improve the distribution of our tests, especially
tasty_available_actions_accepted. However, this does mean that we
discard more frequently. To avoid tests failing due to too many
discards, we reduce the number of trials of tasty_genProg_well_formed
and tasty_tcWholeProg_idempotent.