fix: substTy must not rename binder to target var
When substituting 's' for 'a' in '∀x.t' when 'x' is free in 's' we must
rename the binder to avoid capture. We choose a fresh 'y' and actually
substitute in '∀y.t[x|->y]'. However, it was possible for us to choose
'y' to be 'a', leading to disaster!

(In the presentation above, we would consider substituting 's' for 'a'
in '∀a.t[x|->a]' and conclude that no work needed to be done. In the
actual implementation we actually always go under the binder after we
rename it to avoid capture, leading to '∀a.t[x|->s]'.)