feat: rework EvalFull; now elide lets more agressively
This commit contains two (related) changes. Firstly, it refactors
EvalFull by more fully splitting out noticing redexes from finding the
normal order redex. Secondly we now check whether a 'let' can be elided
(because its bound variable is never used) more agressively, changing
evaluation order slightly (we will now elide a let whenever possible,
whereas previously we may rename some interior binders to avoid capture
for the hypothetical case where we need to substitute underneath them).

To do this we may recompute the free variable set of subterms fairly
often (on every iteration, or possibly more often in the case of nested
lets). This could be optimised in the future. However, it does not seem
to be a problem in practice (for the runtime of our testsuite at least).