feat: Don't take selection input in `getAvailableActions` endpoint
It's unnecessary, since the server already knows the selection. And when we flesh out the API, it will become necessary for clients to keep the server constantly informed of selection changes anyway. For example, the actions API is designed with the assumption that the selection state comes from the backend.

Since the request no longer contains a body, we are able to make it a `GET`.