chore: update eval full tests for agressive let elision
We slightly change three tests
- unit_type_preservation_rename_LAM_regression
- update unit_regression_self_capture_let_let
- tasty_resume_regression
to use the let-bound variable, so that they do not rely on the fact that
we prefer to rename bindings which may capture outer variables instead
of eliding them (when they are unused 'let's).

This is in preparation for eliding such 'let's in the next commit.