feat: logs have trace ids (except db logs/exceptions)
Currently this is somewhat pointless, since we don't actually log
anything other than from the db. However, this is set up so that when we
do, we will automatically attach the correct trace id.

In a follow up commit we will attach the correct trace id to logs
originating in the DB. The difficulty here is that the DB is essentially
a separate service (runs in a separate thread, and serves requests from
all api calls), so we need to attach the trace id at the source, rather
than at a higher level. This means that we will need to include it in
the DB's opQueue.