refactor: move test utilities, prep for test of Eval
We wish to property test the interactive/one-step Eval. We need to use
some generation code that currently lives in Tests.EvalFull, but that
module currently depends on Tests.Eval. This commit moves the
(soon-to-be) common code into Tests.Eval.Utils and Test.Util.

We also generalise the type of 'failWhenSevereLogs', which necessitates
some new type applications at use sites.