test: redexes of Eval are independent
We test that reductions do not interfere with each other, in the sense
that if a term has two subexpressions which are both redexes and we
reduce one, then the other is a redex in the result. This holds unless
the subexpression no longer exists (e.g. it was part of a now-elided
'let'), or it was a no longer required renaming.