feat: EvalFull log CaseRedex invariant failures
These should never happen, since we assume that any term we pass to
EvalFull is type correct. (However, we do not enforce this for our
testsuite, which can lead to confusing test failures if these invariants
are broken.)

Note that we slightly change 'extract', so we only warn if no branch
found once instantiateCon succeeds. Consider the empty case expression
  case C : ? of {}
which is type correct (the scrutinee has hole type, so we expect no
branches), but there is no branch for `C`. This is a perfectly valid
program, which will not reduce. We do not want to log any warning here.