refactor: EvalFull.CaseRedex do not zip Expr and Type
This has two reasons. Firstly for consistency with keeping the binder
names split out. Secondly (and the reason we do not also zip with the
binder names) we would like to add logging when these lists differ in
length (it should never happen if the input has passed typechecking, but
at the very least our testsuite does not typecheck the inputs, which
could potentially lead to confusing test failures). Because the list of
type annotations is `forall m . MonadFresh NameCounter m => ...`, we
cannot do this unless we are in such a context, and we only want to do
this when running a specific redex, rather than contemplating whether an
expression is a redex in the first place.