refactor: EvalFull case reduction compute type with all type arguments
When the input is type-correct, this change will not make any
difference, as a constructor in scrutinee position must be exactly
saturated for it to inhabit an AST. Thus 'tyargs' and
'astTypeDefParameters tydef' have the same length.

For ill-typed inputs, this could change evaluation, but those are not
supported. However, we will shortly log if we detect this particular
invariant being flouted, and this will be easier if we do not drop extra
type arguments.