feat: merge evaluators - finding redexes
The single-step evaluator now finds redexes via a thin wrapper around
the logic that the run-to-normal-form evaluator is built upon.

We update tests since we have changed the notion of reduction.

Two tests (unit_redexes_letrec_{app,APP}_1) fail, due to uncovering a bug.
This will be fixed in a subsequent commit.