test: remove Eval tests for Eval-specific features
We remove three types of test.

Firstly, beta reductions with name clashes.
We currently have special treatment of terms such as '(λx. _) x', by
renaming the lambda-bound variable and reducing the beta in one step.
When the evaluators are merged we will not have this special handling.

Secondly, pushing applications into letrecs.
These have a special rule currently which will not be supported when we
merge the evaluators. The rules for 'letrec' will change so they are
handled uniformly with 'let', and thus we can simply inline and elide
the 'letrec' to expose the beta redex.

Thirdly, ill-typed case expressions.
Currently these throw custom errors, but when we merge the evaluators
they will no longer. This should not matter since it is expected that
only well-typed terms will be evaluated.