feat!: merge evaluators - reducing expressions
The single-step evaluator now reduces expressions via a thin wrapper
around the logic that the run-to-normal-form evaluator is built upon.
This completes the merging of the two evaluators.

Note that the set of reduction rules has changed, since the two
evaluators were not previously consistent in this regard.  There are
lots of small changes in the testsuite as knock-on effects. Mostly
- beta redexes need annotations
- name-clash is handled uniformly now
- no special rule for pushing app into letrec
- we annotate things in the output more
There is also a small regression with the fresh names generated (in the
self-capturing-lets tests).  Previously there was some custom code (in
'makeSafeLetBinding') that found a unique numerical suffix when
renaming, but now we simply call 'freshName'.

We don't detect as many error cases in evaluation. We will now say
'NotRedex' where previously we would say 'BadLambdaAnnotation' etc.
This should not be noticable when reducing well-typed expressions.

We add some more fields to the 'Redex' records to enable computation of
'EvalDetail's.

BREAKING CHANGE: this adds a summand to EvalDetail, which is exposed via
the /api/sessions/:sessionId/eval-step endpoint. (This endpoint is
currently only exposed via the full-featured API and not the simplified
OpenAPI-described one.)