fix: viewCaseRedex must only accept one layer of annotation
Otherwise, the following two expressions would be treated the same:
  case ((C : ?) : T) of ...
and
  case (C : T) of ...
even if (only) the second is ill-typed. This leads, at the very least,
to situations where EvalFull will log warnings when it should quietly
say 'NotRedex' (for instance, if 'C' is not a constructor of the type
'T').

This fixes a regression introduced in
a2a3a5c350e4d65c0fdc7ca517127ee1ca733040