feat: Type holes can act as higher-kinded foralls
The status quo is that type holes can act only as foralls whose bound
variable has kind Type. This means that there are some valid judgements
which become false when replacing a forall by a type hole. For instance
∀a:Type -> Type . a Bool ∋ Λa. ? : a Bool   typechecks, but
                       ? ∋ Λa. ? : a Bool   does not,
since the type can only act as  ∀a:Type. ?, but the RHS requires 'a' to
be higher-kinded.