feat: evaluate `let`s by "pushing down", with opt TODO: squash this with next commit!
This can be seen as an explicit-substitution style operational
semantics.

This change in approach means that our evaluation rules are much more
local, and will enable some future simplifications. In particular we
will avoid using the `Accum` machinery, simplify the eval `Cxt` and not
have to have a separate eval mode for substitutions.

We temporarily disable EvalFull.unit_8. It does pass (note that the step
count is bumped), but takes a long time (almost 10 minutes on my
machine!). This will be massively improved shortly, by being more
aggressive in eliding pointless `let`s.

Signed-off-by: Ben Price <ben@hackworthltd.com>

pushMulti opt