debugging... I see the problem:
before these changes, at this point in the test (note bits are commented
out) the edits to bar (with SH on) have made 'main' have the form
  case <something of type THole> of ...
Previously to the changes in this branch, smartholes would have kicked
in and removed the branches, as the scrutinee is a hole. However, now it
does not. Thus changing the type of main breaks things since the
branches now do not TC