refactor: Create a new primer-testlib package.
We have numerous testing-related utility functions that are repeated
in multiple packages' test suites. This "testlib" package will help us
keep them in one place.

Note that this commit simply creates the new library and uses it in
the `primer` test suite; it is not yet used in the `primer-rel8` or
`primer-service` test suites.