feat: Simplify the available actions API
This is all motivated by the need to develop an API for actually applying actions, to be usable by a "dumb" client which lacks deep knowledge of the semantics of types and expressions. The old code was ported from an old PureScript frontend, and turned out not to be fully suitable for this purpose, due to a need to serialize types and expressions, as well as some use of GADTs and functions, which are not serializable.

We treat `ProgAction` as a low-level notion of "action", to be used by the backend only. We introduce a type `Available.Action` for actions which can be triggered from a definition or a node within it (in our primer-app frontend, this corresponds to actions which appear in the "action panel"). This roughly replaces the old `OfferedAction` type, with descriptions and other textual metadata now being left up to the client, analogously to how the tree API now only returns a `NodeFlavor`, allowing more flexibility in terms of rendering, i18n etc. This new type is just a union of two enums, and thus serialization friendly. Functions in `Primer.Action.Available` which returned an `OfferedAction` are modified (heavily) to return these new actions. Future separate APIs, such as for adding definitions and type definitions, renaming modules, or modifying settings like smart holes, will target the remaining constructors of `ProgAction`. This commit includes the mapping from `Available.Action` to a sequence of `ProgAction`s (i.e. `toProgActionNoInput` and `toProgActionInput`), which is adapted from the old implementation, but isn't actually used yet outside of tests.

Available actions are split in to `InputAction` and `NoInputAction` variants, where the former require further input (e.g. a constructor or variable name) in order to be applied. We provide an API for clients to query what options are available for an `InputAction`. This functionality was not usable prior to this commit, as it went via the "questions" interface, which is not exposed via OpenAPI. Rather than tackling the future of the questions API here, we merely borrow some code from its implementation, though it's so simple after being inlined and simplified as to not be worth attempting to DRY.

Note that while little of the old implementation remains, changes were made fairly mechanically, rather than ripping up and starting from scratch. So it should be the case that, for any given definition or node, the actual set of actions offered hasn't changed. And while the diff of `primer/test/outputs/available-actions/M.comprehensive` is too large for it to seem worth painstakingly manually checking that the new outputs are totally equivalent, we have verified that none of the lists returned contain duplicate entries and that list lengths are as before (though technically some fairly simple refactors were performed since this was last checked, and the bespoke code used to do so is no longer applicable). We also still have the "available actions accepted" property test, which provides reasonable assurance that all offered actions can be run without triggering errors.

The simplicity of this API has some obvious minor issues:
- Some type safety is lost, compared to the old API, in terms of separating actions which apply to definitions versus types versus expressions. We could look to separate these, with three different API endpoints, each with their own action types, and input/no-input variants. Any approach based on GADTs or witnesses is likely to be incompatible with wanting to easily support dumb clients. Anyway, our tests give us some confidence that offered actions will always be safe to apply.
- There is some redundancy in the client sending back a full name received from the backend for the input to an action e.g. for using a variable, rather than the index in the offered list. Though using an index would require the backend to remember some state, which we'd prefer to avoid. Names are likely to be short enough in practice that this isn't really an issue.
- Given that a client is likely to often query the input options available for an action shortly after querying the list of actions available, the backend may often have to perform two expensive find operations for the same node in quick succession, due to the separate API calls for "available" and "options". Indeed, once we add an API for applying actions, this will become _three_ such calls. We could look in to caching the results.