fix: type preservation for evaluating case redex
In the (somewhat edge) case where we scrutinise an annotated known
(polymorphic) constructor, where the types from the type applications
and the annotation differ (necessarily the two types are consistent, but
they may differ in where holes occur), we can have reduction steps that
do not preserve well-typed-ness. See the comments and the added test
case for an example.

This happens because we replace the case expression with the RHS of the
appropriate branch, wrapped in one let binding per pattern variable,
binding that variable to the appropriate argument to the constructor.
These bindings are annotated (because of bidirectionality), but for the
binding to be well-typed, we needed to have annotated with the type from
the application, and for the RHS to be well-typed we needed the one from
the original annotation on the scrutinee. However, we always annotated
with only the one from the scrutinee, leading to reducts which were
ill-typed. We now annotate twice, the inner one arising from the type
applications and the outer from the original annotation. (But in the
case where these are alpha-equivalent, we only use one annotation.)