test: Report benchmark R² in tooltip rather than as separate graph.
The github-action-benchmark action can be configured to warn the
maintainers if the delta between one run and the next is over a
certain percentage. The default is 200%, which seems high and we're
likely to choose something lower. But even 200% would certainly be
triggered fairly often for reported R² values, and there's no way to
configure the action to ignore certain values, so this doesn't seem
like it will be very useful.

Instead, we now report the R² in the `extra` property, which will show
up when hovering over a given OLS value. This is less useful IMO than
a running chart of R² measurements, but probably the right tradeoff
until/unless the action gets support for ignoring values, or, even
better, getting support for reporting when R² dip below a certain
threshold (e.g., 90%).