feat!: openapi actions for primitives
We also rename the 'optLocal' helper function to reflect that it is used
for more than local variable names.

Note that we choose to accept strings (rather than a json number) as the
OpenAPI input for insertion of a primitive integer. This is because
- this API is intended for web/browser based clients and native html
  (numeric) input forms always represent their contents by strings.
- we wish to sidestep issues around JavaScript/JSON support for
  arbitrarily-sized integers.
- it eases implementation, since we can reuse the existing 'InputAction'
  and 'Option's.

BREAKING CHANGE: this changes the OpenAPI interface.