feat: Add an OpenAPI endpoint for creating a new definition
It makes sense to have a single top-level endpoint for this. It is fundamentally different to anything in our existing "actions" API, since it is always available and doesn't correspond to a particular selection. And metadata in that API, such as priority order, would be unused or make no sense. The only other `ProgAction` constructors which aren't targeted by anything in `Primer.Action.Available` are related to modules and type defs, and these will eventually have APIs of their own. So `CreateDef` is one-of-a-kind.