fix!: Make delete session API endpoint REST-y.
What was previously a new `delete-session` endpoint with a `DELETE`
verb that took a body parameter (the UUID of the session to delete),
is now just a `DELETE` verb on the `/sessions/:sid` endpoint. This new
scheme is more REST-y. Some clients (e.g., Axios) have traditionally
had issues supporting `DELETE` with a body, and it appears to be
unconventional outside of mass deletion requests, in any case.

BREAKING CHANGE: this changes both the Servant and OpenAPI APIs,
though we don't currently have any external clients that use the
now-removed `delete-session` endpoint.