chore: add some Read instances
We add these instances as setup for "replay fixtures", where we wish to
record a list of actions in a file and replay these as a benchmark. The
strategy is simple: we already log API calls, so we can extract actions
from primer's log; we simply need to be able to 'Read' them in from a
file.