chore: Switch to haskell.nix nixpkgs-2211 pin, re-enable Docker test.
Turns out the haskell.nix `nixpkgs-unstable` pin is quite old -- older
than its `nixpkgs-2211` pin, anyway -- and I believe the
`nixpkgs-unstable` pin pre-dates the NixOS test framework changes that
caused the docker-image.nix test to break. The utility functions we
use to include these NixOS tests in our flake depend on the new NixOS
test framework behavior, so the theory here is that this is an
incompatibility caused by the ancient `nixpkgs-unstable` pin.

Let's switch to the `nixpkgs-2211` pin and see if this makes that test
work again.