chore: import haskell-dangerous-functions hlint warnings
We follow haskell-dangerous-functions and forbid many functions that it
is better not to use. Note that we have some differing opinions, so we
locally have removed some restrictions.

We also comment out three (sets of) rules, which we would like to
reenable in the future, and add a comment that 'lens' is forbidden in
favour of 'optics' (whereas upstream forbids it as "debatable, should
require a good justification").

One of the functions forbidden is 'maximum', since it is partial (when
given the empty list). In the two places we use it we add an annotation
to ignore the warning since we are sure it will never be called with the
empty list (and there is a pre-existing comment to this effect in one of
these locations).