fix: forkIO -> withAsync, testsuite leaks threads
Follow best practices and use the 'async' library, which is a thin layer
over the concurrency primitives which add some safety and robustness.

This change will stop the testsuite from leaking the 'forkIO'-spawned
threads, since 'withAsync' will ensure the spawned thread (first
argument) is killed when the main action (second argument) returns (or
throws an exception).

Now that 'forkIO' does not occur in the codebase, we can easily add a
hlint rule to ensure we don't unthinkingly use it in the future.