fix: Resurrect the docker-image NixOS test.
Note: for some reason, we need to use `hostPkgs` in the NixOS config
wherever we want to use a Primer package (e.g., the Docker image), for
some reason I don't understand. It seems to have something to do with
the haskell.nix overlay, because other packages that are in our
overlay, but that don't rely on haskell.nix (e.g., `primer-sqitch`),
work fine from the `pkgs` set.