ci!: Make a note of a benchmarking change, add table to README.
BREAKING CHANGE: with this commit, we have made a change to our
benchmarking hardware. See the comment in the new table in the README
for details. Previous benchmark results are no longer relevant to
subsequent results. Please ignore any protestations from our GitHub
benchmark action.

Going forward, we'll endeavor to document benchmarking changes in the
README, which will make it easier to compare results.