ci!: Add new benchmark checks to Buildkite pipeline.
Note: this changes the names of the "required" Hydra CI jobset, and
adds a new required benchmark jobset. This change relies on
https://github.com/hackworthltd/hackworth-ops/commit/44166c22a121288be76fefe1c354e56cb79621cd
having been deployed, which synchronizes the Primer GitHub repo's
required checks. (It has been deployed at the time of writing.)

It also relies on an updated version of our nix-buildkite plugin fork,
which supports the new required job names, and can set agent tags on
steps generated by the plugin, so that we can ensure that benchmarking
jobs run on agents tagged with "private=benchmarking".

BREAKING CHANGE: this change requires that all existing PRs be
rebased, so that they can use the new required job names.