feat: Add an impurity to benchmarks to allow forced re-runs.
Because we format this impurity as a string containing an ISO date
plus version (e.g., "20230130.01"), we can also use it to keep track
of which environment was used to run a particular benchmark run. For
each benchmark result that we store in the Nix store, we also now
create a `lastEnvChange` file, in addition to the benchmark results.