feat: separate replay logger
To record fixtures to replay (for debugging or benchmarking), we
essentially want detailed logs. Currently our standard logs are verbose
enough to enable this, but this will not be true in the future. We
decouple these two uses, so (in future work) we can modify what we log
by default without breaking the workflow of recording fixtures.