feat: test results of benchmarks
This ensures that we are benchmarking what we think we are benchmarking.

The approach here is to more abstractly describe our benchmarks and
assertions about their outputs. We then interpret this into both a
'criterion' based benchmark suite and a 'tasty' based testsuite.  The
tests and benchmarks are independent (i.e. we don't measure the time to
run the assertion in the benchmark stats).