chore(benchmarks): Add cpuTime benchmarks to the benchmark output.
Turns out Criterion has supported this feature all along, and I just
wasn't aware: we can now output benchmark times as measured by the
time the benchmarking process spent running on the CPU, as opposed to
the wall clock time. CPU times should be much more stable.

Note that we still output the wall clock times, as in a perfect
benchmarking environment, that is the true performance metric.
Parallelized algorithms and concurrent threads, for example, will
report the sum of all CPU times across all cores, and would not
reflect the fact that, in real time, the computations had overlapped.