refactor!: Use stronger types for node flavors
A node's flavor determines what sort of body it has, and we now capture this invariant. This will allow clients to avoid some ad-hoc error handling for cases which don't make sense, where we already "knew" that, for example, an application node wouldn't have a body, or a box body had to be a pattern, but we couldn't convince the type-checker of this.