feat: show patterns in saturated form
This is a start towards having constructors be fully-saturated. Once
constructors are in saturated form (without storing their indices/type
applications) also, there will be much more symmetry between
constructions and pattern matches (i.e. the tree structure of creating
`C s t` and a pattern match `case e of C x y -> ...` will be the
same).

This commit changes the data we emit via the OpenAPI endpoints.
However, it obeys the same schema, and is thus not a breaking change.