feat!: constructors can directly store (type and term) arguments
This is a small step towards having constructors always being
fully-saturated. We now have the option for constructors to be
partially saturated: they can store some arguments, and receive others
by an `App` or `APP`. In the future, the typechecker will enforce full
saturation; but before that we need to change all usages to make
fully-saturated constructors.

(To clarify the terminology, consider a	`Cons` constructor, which has
one type argument and two term arguments.
The following Primer `Expr`s (eliding metadata) are completely
unsaturated (i.e. the old status quo):
- `Con "Cons" [] []`
- `APP (Con "Cons" [] []) _`
- `App (APP (Con "Cons" [] []) _) _`
- `App (App (APP (Con "Cons" [] []) _) _) _`
Note that `Con "Cons" [] []` after this commit corresponds to
`Con "Cons"` before this commit.

The following is fully saturated (i.e. will be the only valid
occurrence in the future)
- `Con "Cons" [_] [_,_]`

The following are partially-saturated
- `Con "Cons" [_] []`
- `App (Con "Cons" [_] []) _`
- `App (App (Con "Cons" [_] []) _) _`
- `Con "Cons" [_] [_]`
- `App (Con "Cons" [_] [_]) _`

In particular, note that "saturated" refers to how many arguments the
constructor is directly applied to, in the sense of being contained in
the lists within the `Con` node of `Expr`. Thus we (are starting to)
treat the arguments of constructors differently to the arguments of
functions.)

There are some slightly awkward edges currently, due to not having
finished our transition from unsaturated to fully saturated. For
partially-saturated constructors (which will be rejected by the
typechecker in the future), we have that
- `typesInExpr` can focus on all type arguments of constructors
- `focusType` will still ignore type arguments of constructors
- the new `focusConTypes` will look at type arguments of constructors
  where necessary (in `Eval.NormalOrder.foldMapExpr`,
  `Eval.NormalOrder.findRedex` and `Zipper.focusOn`)

There are some `TODO` notes documenting some of the work we still have
to do, and some of the slight strangeness around this halfway point.
These will be dealt with in subsequent commits.

BREAKING CHANGE: this commit changes the `Expr` ADT, which will break
programs created before this commit (i.e. we cannot load databases
created by older versions of primer).