feat: Improve `primer-sqitch` script, better testing.
We make the following changes:

* `primer-sqitch` now works with SQLite databases, in addition to the
existing PostgreSQL support.

* We test its SQLite support and add the test to the flake's `checks`
attribute, so that CI will require it to pass.

* Both the SQLite and PostgreSQL now test a deployment, a revert, and
another deploy, to ensure all of these operations work as expected (on
a new/empty database, anyway).

* We eliminate the `db-init` NixOS test and roll its functionality
into the `db-tests` NixOS test (now called `db-postgresql-tests`), to
be a bit more efficient in CI, as the NixOS tests are relatively
expensive.